+++ src/components/Widgets/Posts/usePosts.ts
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import { Post } from '@typings/posts/post';
 import { getPosts } from '@server/posts/getPosts';
 
-export default function usePosts(limit: number, category: string) {
+export default function usePosts(limit: number, category?: number) {
   const [posts, setPosts] = useState<Post[]>([]);
 
   useEffect(() => {