+++ src/server/shop/getCustomer.ts
@@ -1,4 +1,6 @@
-export async function getCustomer(field, value) {
+import { Customer } from '@typings/shop/customer';
+
+export async function getCustomer<K extends 'id' | 'email'>(field: K, value: Customer[K]): Promise<Customer | undefined> {
   try {
     const res = await fetch('/data/shop-customers.json');
 
@@ -6,10 +8,8 @@ export async function getCustomer(field, value) {
       throw new Error(`Failed to fetch shop-customers.json: ${res.status}`);
     }
 
-    const customers = await res.json();
+    const customers: Customer[] = await res.json();
     const customer = customers.find((customer) => customer[field] === value);
-
-    if (!customer) return;
 
     return customer;
   } catch (error) {
@@ -18,10 +18,10 @@ export async function getCustomer(field, value) {
   }
 }
 
-export function getCustomerById(id) {
+export function getCustomerById(id: number) {
   return getCustomer('id', id);
 }
 
-export function getCustomerByEmail(email) {
+export function getCustomerByEmail(email: string) {
   return getCustomer('email', email);
 }