+++ src/server/shop/getShippingOptions.ts
@@ -1,4 +1,6 @@
-export async function getShippingOptions() {
+import { ShippingOption } from '@typings/shop/shippingOption';
+
+export async function getShippingOptions(): Promise<ShippingOption[] | undefined> {
   try {
     const res = await fetch('/data/shop-shipping.json');
 
@@ -6,9 +8,7 @@ export async function getShippingOptions() {
       throw new Error(`Failed to fetch shop-shipping.json: ${res.status}`);
     }
 
-    const shippingOptions = await res.json();
-
-    if (!shippingOptions) return;
+    const shippingOptions: ShippingOption[] = await res.json();
 
     return shippingOptions;
   } catch (error) {