+++ CHANGELOG.md
@@ -27,6 +27,7 @@ All notable changes to this project will be documented in this file.
 #### Components
 
 - Posts widget category incorrect type
+- Products widget category incorrect type
 
 ## [0.1.0] - 2025-11-25